isl_map_coalesce: only harmonize divs if equality constraints are different
Integer division harmonization was originally introduced in commit
8eb322e (isl_map_coalesce: harmonize integer divisions that differ by
a constant, Mon Oct 13 22:54:29 2014 +0200) to handle integer divisions
that only differ by an integer shift in the constant term.
This was recently extended to take into account the equality constraints.
The motivation for the original harmonization was however removed
by the previous commit, since such shifts should have been removed
as a special case of the reduction of the coefficients of the integer
divisions.  The extension may still be useful however, but only
when the two basic maps do not have identical equality constraints.
Only perform the harmonization when those equality constraints are different.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>