isl_printer_print_ast_expr: support an isl format
Ever since its introduction, isl_printer_print_ast_expr has printed
its argument in C format, irrespective of the printer output format.
This was arguably a mistake.  The isl format of practically all
other isl objects allow the object to be parsed from the output,
which is not possible from this C format if only because several
isl operation types are mapped to the same C operator.
This will become even more problematic when additional information
is added to the expressions that does not affect a straightforward
C representation.

Introduce a tailored isl format for isl_ast_expr objects and
only use the C format when explicitly requested.
In theory, this breaks backward compatibility, but the damage
should be relatively minor in practice.  Users that need
a C representation should have been explicitly selecting the C format
already.

In analogy with isl_schedule objects, the isl format is in the form
of a YAML object.  This should make it relatively easy to implement
parsing of isl_ast_expr objects from their textual representation.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>