python interface: expose construction of AST node from union map
This change makes basic AST node generation available in the python bindings.
We currently only expose AST generation from a schedule map, as it is
the easiest way to showcase the code generation facilities of isl. In a
subsequent step, we may want to expose the schedule tree based code
generation as well.

isl_ast_node objects use the new YAML based formatting for their
__str__ and __repr__ representation.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>