isl_tab_pip.c: isl_partial_sol: use isl_multi_aff instead of isl_mat
isl_multi_aff objects are easier to work with and less error-prone because,
once created, they are less sensitive to changes in the representation
of the domain.  An isl_multi_aff representation will also make it possible
(in a future commit) to check for equal partial solutions that are
expressed differently.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>