isl_map_coalesce: allow separating equality constraints in extension
Before this commit, isl_map_coalesce would break early as soon as
there is any equality constraint that separates one basic map
from the other.  However, if one basic map is an extension of
another, then simplification of the extension may have resulted
in separating equality constraints.  Furthermore, the check
that determines whether a basic map is an extension of another
only relies on the status information from this other basic map.

Delay the check for separating equality constraints until
the status of every constraint has been determined and
then check for one being an extension of the other
in case of the right pattern.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>