isl_stream_read_pw_multi_aff: read isl_pw_multi_aff directly
Do so instead of first reading in an isl_map or isl_set and
then converting that to an isl_pw_multi_aff.
The indirect way could result in a representation that is
quite different from the input, causing needless confusion.
It also failed to properly read expressions involving NaNs.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>