isl_tab_pip.c: set free callback before any other fields
This ensures that isl_sol_free (which calls this callback) can safely
be called at any time after this initialization.
The alternative would be to make the call to the callback conditional
on the free callback having been set, but doing it this way ensures
that no memory is lost in case some private fields may have
been set before the free callback is set.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>