isl_basic_{set,map}_plain_cmp: drop const qualifier
The const qualifier is not used very frequently in isl because
many functions that do not change the meaning of their arguments
might still change their internal representations.
Even though isl_basic_map_plain_cmp does not do so at the moment,
it might very well do so in future.
Remove the const qualifier as it only causes some clutter
in the current code base.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>