isl_map_simplify.c: ok_to_eliminate_div: also investigate after unknown div
The function ok_to_eliminate_div checks that there is no integer division
that involves the integer division that is going to be eliminated and
that is positioned before the last integer division involved
in the equality constraint that will be used to perform the elimination.
The original code, which has been in there since 2973320 (isl_map_simplify:
don't remove any div definitions if all divs are known,
Wed Apr 7 15:59:02 2010 +0200), would stop investigating when it runs
into an unknown existentially quantified variable.
Perhaps the reasoning was that there would be no explicit integer divisions
after an unknown existentially quantified variable, but that may still
happen in the current code base if an explicit integer division
involves an unknown existentially quantified variable.
It is not clear if this can cause any problems in practice,
but it seems safer to continue looking at the integer divisions
even after running into an unknown existentially quantified variable.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>