isl_basic_map_drop_redundant_divs_ineq: extract out any_div_involves_div
This makes it easier to apply the change in the next commit.
There is already one very minor change in this commit
in that the extracted function will only start looking
at the next local variable, while the original code
started looking at the current local variable,
but the explicit representation of a local variable
could never involve itself, so this should not make a difference.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>