isl_map_coalesce: add check that basic map and tableau have same constraints
The function expand_tab expands the tableau representation to match
the basic map representation, which has been extended with extra
variables and constraints since the tableau representation was
originally derived from the basic map.
Check that no constraints were added to the tableau in between
such that the two representations would still match after
adding the extra constraints in the basic map to the tableau.
That is, check that the situation that could occur before the previous
commit, where the original basic map having implicit equality constraints
caused additional constraints in the tableau, no longer occurs.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>