isl_tab: keep track of location of inserted integer division in undo stack
When an integer division is added to the basic map representation
of an isl_tab, it is currently always added at the end.
However, this will be changed in an upcoming commit.
That is, integer divisions may also get added at other positions.
Store the position of the added integer division in the undo stack and
take this position into account when dropping the integer division on undo.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>