isl_basic_map_add_div: explicitly zero out coefficient of new integer division
isl_basic_map_add_div takes a description of the new integer division
in terms of the original variables.  In particular, the description
does not reference the newly added integer division itself.
When initializing the integer division from this description,
the coefficient of the integer division itself is therefore not explicitly set.
isl_basic_map_alloc_div does take care of initializing this coefficient
to zero, but it should be more future-proof to initialize the coefficient
explicitly inside isl_basic_map_add_div.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>