isl_output.c: call print_set_c with proper space
Since 3280c05 (make isl_pw_* object live in a map space,
Tue Aug 30 16:47:59 2011 +0200), print_set_c was getting called
with the space of a map.  Call it with the domain of this space instead.
Clarify the meaning to the "space" argument of some functions
called with the same domain space to support the correctness
of this change.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>