deprecate isl_set_align_divs and isl_map_align_divs
These functions expose too much of the internal representation and
will be removed in a future release.  For now at least discourage
new users of these functions.
It is not clear if there are any outside users.  If there are,
then perhaps it could be useful to add an isl_map_lift.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>