privately export isl_map_is_params
This function will be used to determine whether the explicit domain
of an isl_multi_pw_aff is a parameter set.
Even though this domain is an isl_set, an isl_set
is currently defined internally to be equal to an isl_map,
resulting in isl_map_is_params getting called when
the function name is derived from the type.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>