isl_ast_build_expr.c: extract_term_and_mod: use term parameter
The parameter 'term' was unused and even violated the __isl_take annotation by
not free'ing the reference. Replace 'data->div' with a use of the parameter.

Since both calls to extract_term_and_mod pass 'data->div' with its reference
counter increased by one as 'term' argument, it just happened to work out and
there is no change in behaviour.

Signed-off-by: Michael Kruse <isl@meinersbur.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>