isl_tab_pip.c: set_from_affine_matrix: check that zero columns are indeed zero
The function sol_pma_add, which calls set_from_affine_matrix, assumes
that the final columns of the matrix (those without a corresponding
variable in the local space) are zero.  Check that this is indeed the case.
Without the fix in the previous commit, this check would trigger
on the test case added in the same commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>