isl_basic_set_multiplicative_call: check for empty unbounded inputs
A rationally unbounded set may still be void of any integer points.
isl_basic_set_multiplicative_call would produce an infinite polynomial
on the parameter domain of this empty input.  While this is technically
correct, it is somewhat confusing.  Check whether the input is empty
such that a zero polynomial is returned in case of empty input instead.

The explicit check for empty domains also allows the subsequent
projection onto the parameter space to drop irrelevant constraints.
In particular, isl_basic_map_project_out only drops (seemingly)
irrelevant constraints if the input is known to be non-empty
as otherwise the "irrelevant" constraints may actually be the ones
that cause the input to be empty.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>