isl_basic_set_multiplicative_call: detect redundant constraints and equalities
Detecting redundant constraints and/or equality constraints
can be relatively expensive, but this may result
in lower-dimensional factors and breaking down a set
into lower-dimensional factors is the whole point of
isl_basic_set_multiplicative_call.
Still, postpone the detection until the factors
of the factorization that can be read off without the explicit detection
since these may already be of lower dimension than the input set.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>