isl_basic_set_compute_vertices: drop vertices with lower-dimensional domain
During the chamber computation, the activity domains are intersected
until a lower-dimensional result would be obtained.  However, the
detection of such lower-dimensional results only takes into account
the inequality constraints of the activity domains, tacitly assuming
that these activity domains do not involve any equality constraints.
In degenerate cases, vertices with lower-dimensional activity domains
do occasionally get generated, resulting in the corresponding vertex
incorrectly being added to one or more chambers.

One solution would be to take into account the equality constraints
during the detection of such lower-dimensional intersections,
but a lower-dimensional activity domain would always result
in a lower-dimensional intersection and would therefore not
appear in any chamber.
It is therefore more appropriate to not even add a vertex with
a lower-dimensional activity domain to the list of vertices.

Without this fix, an extra "[n, n, n]" vertex would be generated
for the new test case.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>