isl_map_convex_hull: use consistent space when lineality space is detected
Internally, the space of the input to isl_map_convex_hull is ignored and
the space of the result of the internal computation may be slightly
different from the input to this internal computation.  This is usually
not a problem because the space of this internal result is replaced by
that of the input of isl_map_convex_hull.
However, when a lineality space is detected during the incremental
convex hull computation, a partial result is combined with a partial input and
in this case a consistent space should be used to prevent
isl_set_add_basic_set from producing an error message.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>