isl_convex_hull.c: uset_convex_hull_unbounded: operate on isl_basic_set_list
Using an isl_basic_set_list means that the functions
isl_set_copy_basic_set and isl_set_drop_basic_set no longer need
to be used.  The latter are two internal functions that are
too closely tied to the internal representation, even for
internal use.  Since these function will be marked deprecated
in the next commit, remove their only use first.

Compared to the original implementation, the new implementation
additionally assumes that the input consists of at least two disjuncts,
but this is always the case.  Inputs with fewer disjuncts are handled
early on and do not reach this stage.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>