isl_ast_build_expr_from_pw_aff: do not include rational expressions in min/max
The input to isl_ast_build_expr_from_pw_aff is assumed to only yield
integer values on its definition domain, but it may involve rational
expressions that are only guaranteed to yield integer values
on the corresponding subpart of the definition domain.
Combining such rational expressions into a min/max expression will
cause the expression to be evaluated on a larger part of the definition
domain where the expression is not guaranteed to yield an integer value.
Refrain from including rational expressions in a min/max expression
to avoid this issue.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>