python: derive all classes from 'object'
This increases consistency with Python 3. Specifically, for classes
derived from object, Python 2 uses the same method resolution algorithm
that is used by Python 3. As the Python 3 algorithm is more strict,
ensuring the use of this algorithm already in Python 2 will make sure
Python 3 support is not regressed by accident by creating a class
hierarchy that does not provide monotonicity.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>