isl_polynomial.c: reduce_divs: do not assume input is a monomial
Even though the original intention may have been to only call
reduce_divs on a monomial, there is no guarantee that some
of the intermediate functions that end up calling reduce_divs
aren't called on other expressions.
Furthermore, the recursive call inside reduce_divs itself
is not guaranteed to have a monomial argument.
Fix these potential issues by not assuming that the input is a monomial.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>