isl_vertices.c: tab_for_shifted_cone: allocate room for equality constraints
The function tab_for_shifted_cone constructs a tableau from
both the equality and the inequality constraints of a basic set,
but it would only allocate room for the inequality constraints.
The function may not actually get called on a basic set
with equality constraints in the current code base, but
it should handle them consistently.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>