Update imath for MSVC compile fix
Commit 01abae0 (Added big-endian support to mpz_import and mpz_export
in gmp_compat.c) of the imath submodule introduces uses of the type ssize_t,
which isl included in commit
bda23e (Update imath to include big-endian support for mpz_import/export).

The type ssize_t belongs to the POSIX standard and may therefore not exist
in non-POSIX environments.  The update adds a typedef ssize_t if compiled
by the Visual C++ compiler.

Signed-off-by: Michael Kruse <isl@meinersbur.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>