isl_aff_add: normalize result
If the addition results in integer divisions with a zero coefficient,
then the result gets printed in the same way as the equivalent
affine expression that does not reference these integer divisions,
but they would not be considered to be obviously equal to each other.
Normalize the result of the addition such that at least some (the final)
of these zero coefficient integer divisions get removed.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>