isl_map_get_space: extract out isl_map_peek_space
This function will be used in places where the space is only needed
temporarily, without exposing too much of the internal representation.
That is, this function can be used instead of just accessing
the "dim" field directly.
The argument is qualified as "const" so that it can also be
used by isl_map_n_* in a future commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>