extract out shared isl_multi_*_{take,restore}_space
This reduces the dependence on the internal representation.

Note that in the case of isl_multi_*_flatten_range,
it is now isl_multi_*_restore_space that is responsible
for calling isl_multi_*_cow when needed.

Also note that even though a call to isl_multi_*_take_space
always needs to be followed by a corresponding call
to isl_multi_*_restore_space, it is perfectly fine
to call isl_multi_*_restore_space without a preceding call
to isl_multi_*_take_space.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>