isl_basic_map_reset_space: extract out isl_space_has_equal_ids
The function isl_space_match will be deprecated in the next commit.
Its last remaining use was inside isl_basic_map_reset_space.
Replace the calls to isl_space_match by a call to isl_space_has_equal_ids,
which can be implemented in terms of the internal "match" function.
As an added benefit, the new code should also be easier to understand.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>