isl_access_info_compute_flow: support "kill" accesses
Kills have the same effect as must-sources in that they
block any dependence between an earlier source and a later source,
but they do not by themselves result in dependences.
Kills are useful in cases where only the dependence blocking effect
is required.

It was already possible for the user to add the kills
to the must-sources and then to remove the resulting
spurious dependences from the result, but this is not
very elegant.
Add direct support for kills.

In fact, make kills the primary interface, with must-sources
considered to be both may-sources and kills since the effect
of kills is actually easier to explain than that of must-sources.

The initial implementation does exactly do the above treating
kills as must-sources and then removing spurious dependences,
but this could be improved in the future.
For example, the pure kills (those that do not overlap
with the actual must-sources) could be explicitly marked
during the computation such that the resulting dependences
would not even have to be added to the results in the first place.
It would also be possible to not even compute may-dependences
derived from kills.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>