isl_union_map: use tuple hash in hash table
This makes the ordering of the maps within a union map
independent of the parameters.
This avoids a source of confusion where some operations
would return different results depending on which parameters
appear in the internal representation.
Since this commit itself changes the hashing function,
it also changes the output of some operations.

Since all the maps in the union map have their parameters
aligned with the union map, there is no need to check
for equal parameters.  In any case, any union map
can only contain a single map in a given tuple space
even if this alignment were not performed.

An upcoming commit will change the comparison of entries
with the same hash value to ignore parameters as well.
Together, these two changes ensure that look-ups can be performed
without first aligning parameters.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>