isl_union_map.c: cond_un_op: unify callbacks
The callers of cond_un_op have callbacks of one of two types,
either they simply apply a map-function to the entry and collect
the result or they additionally first check whether the map-function
can be applied.
Replace these callbacks by a single callback that performs
these operations based on input set by the caller of cond_un_op,
in particular, the function that should be called on each map and
the function that is used to filter the maps (if any).

This removes some code duplication and, with some further modifications,
will allow inplace and un_op to be implemented in terms of cond_un_op.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>