isl_coalesce.c: is_adj_eq_extension: check for redundant constraints
If any of the relaxed constraints turn out to be redundant
then isl_tab_select_facet cannot be called on that constraint.
The failure to check for redundant constraints beforehand
would result in a failed assertion inside isl_tab_select_facet
on at least one instance.  Reproducing this problem is tricky
however so no test case is included in this commit.
While it may be possible to handle such constraints, it is not
clear if this would ever be useful, so simply bail out instead for now.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>