isl_scheduler.c: count_carry_edges: only count validity edges
Commit d07d9a4 (isl_union_set_compute_schedule: only carry validity
dependences, Mon Mar 26 12:55:35 2012 +0200) modified setup_carry_lp
to only attempt to carry validity dependences.  However, it failed
to update the way the number of edges is counted.
This can result in extra variables in the LP formulation, corresponding
to the non-validity edges, without any constraints.
This should in itself be fairly harmless, but it does mean
that these extra edges can always be "carried", which in turn
means that the procedure will fail to detect cases where
no actual edges have been carried.
Only count validity edges to avoid this issue.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>