isl_scheduler.c: compute_schedule_finish_band: split along SCCs before carrying
compute_schedule_finish_band resorts to carrying dependences using
Feautrier if all else fails.  If there is more than one SCC left,
then whatever is causing the failure to create a new schedule row
may be limited to one or a few SCCs.  It can therefore help
to split the graph into SCCs first.
Note that this can only happen in case of whole-component scheduling and
if no suitable cut could be found.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>