isl_scheduler.c: carry_dependences: try separate components in case of no edges
If there are no edges to be carried, then any attempt to carry edges
is guaranteed to fail since the number of carried edges will always be zero.
If the dependence graph consists of more than one component,
then it may still be possible to compute a schedule on each component
separately, so try that instead.

One of the test cases already has zero (validity) edges, but due
to a mistake in the way edges are counted, pure proximity constraints
are counted as well.  This will be fixed in the next commit,
resulting in a failed test case without the changes in the present commit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>