isl_scheduler.c: fix documentation of inter_dim_map and callers
In particular, the generic form of the valid constraints has separate
c_x and c_y coefficients and the coefficients for the variables of
the two nodes are plugged in for c_x and c_y separately (with the appropriate
sign).  It is not clear why the original documentation tries
to plug in a difference between these coefficients, but the mistake
was already introduced in 1192654 (add isl_union_set_compute_schedule,
Sat Mar 19 14:47:47 2011 +0100).

The documentation of add_inter_proximity_constraints had a further mistake
introduced in 2a66550 (isl_schedule_constraints: add support for conditional
validity constraints, Sat Mar 24 12:09:09 2012 +0100).
In particular, the sign was only changed in part of an equation.
Fix the signs and add a clarification that when local is set, the sign "s"
is equal to 1.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>