Turn python generator into a subclass of generator
We introduce a generator class that carries the printing logic, which
was previously stored together with all other information in the isl
class. Keeping language generic printing functionality in a separate
class and moving language specific code into a subclass allows us to
introduce generators for new languages easily.

This change is large, but mostly mechanical. In generator.cc the prefix
'generator::' is added to turn static functions into class methods. In
python.cc functions are moved into the newly introduced
python_generator. This sometimes requires the introduction of a new
isl_class argument.

The only non-trivial change is the separation of generate_python into
generator::generator() and python_generator::generate(). The former now
constructs the set of isl_classes  which we want to generate bindings
for, whereas the latter prints for each class the corresponding
bindings.

We also move the isl_class::is_static method into generator. We had a
choice to either move extract_type into isl_class or to move is_static
into generator::, if we wanted to avoid cross referencing functionality
from both classes. As is_static is not an inherent property of the class
method, but our choice of how to generate code for a class, it seems
best placed in generator:: alongside similar functions such as
is_callback or is_constructor.

This change does not change current behavior. The generated isl.py files
are identical.

Signed-off-by: Armin Groesslinger <armin.groesslinger@uni-passau.de>
Signed-off-by: Andreas Simbuerger <simbuerg@fim.uni-passau.de>
Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>