isl_basic_map_drop_redundant_divs: drop explicit representation of coalesced div
Div coalescing replaces a pair of existentially quantified variables
a and b such that a only appears separately in two constraints
e <= a <= e + m - 1 and otherwise only appears together with b
in the combination a + m b by a single existentially quantified variable
a + m b.  The constraints that only contains b are furthermore required
to have coefficient 1 or -1 for b.

Div coalescing usually does not get applied to local variables
with an explicit representation because the defining constraints
have coefficients different from 1 or -1.
Those constraints may not be explicitly available, though, and
so it can occur in practice that div coalescing is applied
while b has an explicit representation.
The div coalescing would fail to remove the explicit representation and
it would get attached to the new local variable with value a + m b,
which is necessarily incorrect.
Remove this explicit representation.

It is difficult to construct a minimal test case where
the defining constraints get removed.
Instead, this commit adds an AST generation test case,
which is a reduced version of an AST generation input
where the problem was reported.

Reported-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>