isl_basic_map_from_aff: call gauss on constructed basic map
Basic maps should always be gaussed before returning them to the user
because many functions assume that basic maps have been gaussed.
In the case of isl_basic_map_from_aff, there can only be a single
equality constraint, so isl_basic_map_gauss would not be able
to use any equality constraints to eliminate coefficients,
but it may still change the sign such that the last coefficient
would have a positive sign.
It is not clear if any part of the current code base depends
on this condition, but it seems more prudent to make sure it
holds given the implicit assumption that all basic maps have
been gaussed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>