isl_scheduler.c: count_map_constraints: drop "carry" argument
The "carry" argument was used for counting the number of constraints
for the LP problem constructed by the Feautrier style scheduler.
Since the previous commit, the basic validity edges are first collected and
the number of constraints is derived from those basic maps.
In particular, the call to count_map_constraints has been removed and
count_map_constraints no longer needs to support this use case.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>