isl_tab_basic_set_non_trivial_lexmin: take triviality directions
The original interface takes a sequence of variables, at least
one of which needs to be non-zero.  This forces the triviality
directions to be available directly as (differences of) variables
in the LP problem.
Take a matrix describing triviality directions instead
in order to allow for more freedom in the encoding of the LP problem.

Note that the original implementation took advantage of
the encoding and enforced x_i = x_i_b - x_i_a = 0 by
introducing two equality constraints x_i_b = 0 and x_i_a = 0.
The new interface can no longer assume that all triviality directions
correspond to coordinate directions and therefore uses
a single equality constraint instead.  In case a triviality direction
does in fact correspond to a coordinate direction, this would
be the equality x_i_b - x_i_a = 0.
It is not clear if this could result in a performance regression.
If it does, then this special case of triviality directions
could be detected explicitly and handled as before.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>