scheduler: reverse order of schedule coefficients
In general, if there is a choice between making one of two coefficients
non-zero, the scheduler should pick the first coefficient.
Since a lexicographic minimum is being computed, it should
therefore appear after the second coefficient.
For the Pluto-like scheduler, the coefficients in the ILP problem
are currently transformed coefficients and the search for non-trivial
solutions already favors making making the first of these coefficients
non-zero because the directions of non-triviality correspond
exactly to these coefficients.
However, it is the actual coefficients that should be considered
in order for making non-zero and not the transformed coefficients.
In an upcoming commit, the ILP problem will refer to the actual
coefficients and the directions of non-triviality will continue
to correspond to the transformed coefficients and may therefore
be different from the actual coefficients.
At that point, the order of the coefficients will have an effect
when the directions of non-triviality are not along coordinate
directions.
This commit therefore has limited effect now, but rather prepares
for this change of encoding.

The order of the schedule coefficients for the LP problem
constructed by the Feautrier style scheduler is also changed
for consistency.
Since the optimization is mainly driven by the number of
carried groups of schedule constraints, the order of the coefficients
is usually not very important for the Feautrier style scheduler.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>