isl_scheduler.c: isl_sched_node: only store rows that need to be non-trivial
The inverse of the schedule coefficient transformation matrix
is used to test whether a solution is trivial.
However, only the final rows in this matrix are needed.
Furthermore, the same rows will be reused to enforce non-trivial
solutions in a call to isl_tab_basic_set_non_trivial_lexmin and
for this purpose it will be useful to perform additional transformations
on these rows.
Store those rows directly instead of the entire inverse.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>