scheduler: change order of coefficients in (I)LP problems
In particular, put the constant term last and
the coefficients of the parameters after those of the variables.
The values of the constant term and the coefficients of the parameters
should not take precedence over those of the variables.
Furthermore, a non-parametric constant term is preferred over
one involving parameters.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>