isl_tab_basic_set_non_trivial_lexmin: reinitialize update field after backtrack
While the very initial value of the update field is zero,
it may have been left set during backtracking.
Reset the value when the level is entered again.

This fix is based on code inspection.
It is not clear if the issue would ever trigger in practice and,
even if it does, it should only result in a needless duplication
of some constraints that zero out initial coordinates.

Still, it seems more prudent to perform a proper reinitialization.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>