isl_sched_node: drop cmap field
The cmap field is no longer used, so it can be removed.
The ctrans field is renamed to vmap since it does not make
sense to have a field that represents the transpose of
a field that has been removed.
Similarly, node_update_cmap is renamed to node_update_vmap.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>