isl_printer_print_union_access_info: do not print empty access relations
If an access relation was not specified, then there is no need to print it.
Since the access relations are empty by default, there is also no need
to print empty access relations even if they were explicitly specified.
The sink access relation on the other hand always needs to be printed,
even if it is empty, because it is used to construct
an isl_union_access_info object.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>