add basic isl_union_access_info_compute_flow tests
isl_test.c already contains an isl_union_access_info_compute_flow test,
but it is rather specialized.  It is easier to specify tests
using the isl_flow application.
This infrastructure will then also make it easier to test
an improvement of isl_union_access_info_compute_flow that
will be introduced in an upcoming commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>