isl_union_access_info_compute_flow: handle coscheduled must-sources
isl_union_access_info_compute_flow is defined to look for dependences
from the last must-source preceding a sink source as well as all
intermediate may-sources or simply any preceding may-source
if there is no preceding must-source.
This definition does not explain what happens when there are
multiple must-sources that are scheduled together or
when there are may-sources that are scheduled together
with a must-source.
What would happen in practice is that one of the must-sources
would result in a dependence, while all other coscheduled sources
would simply be ignored.

It can sometimes be useful to allow coscheduled sources,
especially in the high-level interface isl_union_access_info_compute_flow.
Given that the input does not specify whether the must-source
is executed before or after the other sources, the only sensible
result is for any dependence resulting from this must-source
to be considered a may-dependence, while the other sources
should also result in may-dependences.

This commit implements this result by looking for coscheduled
sources after an initial computation of the dependences
derived from a must-source at a given level.
The behavior with respect to sources that are not coscheduled
is not changed by this commit.
The change in behavior is also reflected in the documentation.

While this commit changes the behavior of isl_union_access_info_compute_flow,
it only does so in cases where it was undefined before.
Moreover, the original behavior was unreliable.

The low-level interface is not modified by this commit.
Users of the low-level interface should continue to ensure
that there are no coscheduled sources.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>