add isl_map_reset_equal_dim_space
isl_map_reset_space is currently used in two situations,
to simply change the space of a map without affecting its dimension and
to finish off a transformation that did affect the dimension.
In the second case, the input map is in an invalid state and
the call to isl_map_reset_space restores validity.
In the first case, the input is in a valid state and should
remain in a valid state.
Add a separate isl_map_reset_equal_dim_space for the first case
that checks that the total dimension is not modified.

This function will be used to adjust the space of the explicit domain
of an isl_multi_pw_aff, when support for such explicit domains is added.
It could also be used to replace some of the current calls
to isl_map_reset_space.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>