keep track of domain in 0D isl_multi_pw_aff and isl_multi_union_pw_aff
This ensures that intersect_domain has an effect on objects
of type isl_multi_pw_aff and isl_multi_union_pw_aff.
This in turn makes it possible to round-trip a conversion
from isl_pw_multi_aff to isl_multi_pw_aff and back to isl_pw_multi_aff
for 0D functions.  This truly makes isl_multi_pw_aff a superset of
isl_pw_multi_aff, which it has been claimed to be in the bindings.
This also makes it possible to carry some information
in 0D isl_multi_union_pw_aff objects such that users do not
have to treat this case specially.
In particular, the prefix schedule at a node without an outer band node
now keeps track of the domain, filtered by any outer filter nodes.

The domain is only kept in case of 0D functions since objects
with at least one base expression carry domain information
in the individual base expressions.
The explicit domain gets taken into account on a product
operation such that it wouldn't get lost.

Since the output tuple of a 0D function is empty,
the explicit domain is printed outside of the tuple.
Note that the 0D output space may still have structure,
so it would not be clear where to place the domain
if it were printed inside the tuple.
In case of an isl_multi_union_pw_aff, the tuple is
the outermost construct in the output.
If an explicit domain needs to be printed, the tuple and
the domain are placed inside parentheses to clarify
that the domain belongs to the empty tuple.

It is not immediately obvious whether a pullback of a 0D function
with respect to another function should intersect the domain
with the (shared) domain of the other function, but this is
what this commit does.

This commit attempts to adjust all affected operations,
but it may have missed some.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>