scheduler: fix handling of bounds on schedule coefficients
Commit 0d3d3ddd0a (scheduler: solve for original schedule coefficients,
Wed May 3 14:29:47 2017 +0200) introduced a bug in the derivation
of one of the constraints on the absolute values of the schedule
coefficients from the other inside node_add_coefficient_constraints,
resulting in the bound not necessarily getting imposed.
It also failed to clear the coefficients, which would then
remain in subsequent bounds on coefficients.
On the other hand, the same commit did silently fix a bug
introduced in commit b7885131b2 (isl_scheduler.c: extract out
shared node_var_coef_pos, Sun May 7 16:04:47 2017 +0200) that
would result in one coefficient overwriting the other.

Reported-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Tested-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>